requirements.txt
asgiref==3.8.1
Django==5.2
django-render-block==0.11
django-rest-framework==0.1.0
django-templated-email==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.2
